+++ example/ct_responsive_example/lib/main.dart
@@ -1,0 +1,19 @@
+import 'package:ct_responsive_example/src/orientation.dart';
+import 'package:flutter/material.dart';
+
+main(List<String> args) {
+  runApp(MyApp());
+}
+
+class MyApp extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      home: Scaffold(
+        body: Center(
+          child: OrientationWidget(),
+        ),
+      ),
+    );
+  }
+}