+++ lib/ct_responsive.dart
@@ -1,11 +1,11 @@
 library ct_responsive;
 
-// custom widgets
+// custom widgets to enable responsive patterns, more commonly accessed through corresponding extension functions
 export 'src/custom_widgets/responsive_by_orientation.dart';
 
-// custom values
+// custom values to enable responsive patterns, more commontly accessed through corresponsing extension functions
 export 'src/custom_values/value_by_orientation.dart';
 
-// extension functions
+// extension functions for more concise access to all the custom classes in the package
 export 'src/extension_functions/responsive_extensions.dart';
 export 'src/extension_functions/value_extensions.dart';