+++ test/value_by_orientation_test.dart
@@ -2,34 +2,126 @@ import 'package:ct_responsive/ct_responsive.dart';
 import 'package:ct_test_helpers/ct_test_helpers.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
+import 'package:mockito/mockito.dart';
+
+class DummyContext extends Mock implements BuildContext {}
 
 main() {
-  group("Value Extensions", () {
-    group("when orientation is landscape", () {
-      testWidgets("get landscape value when landscape",
+  final testPortraitSize = Size(320, 800);
+  final testLandscapeSize = Size(800, 320);
+
+  group("Extension Functions", () {
+    group("valueWhenOrientationLandscape", () {
+      testWidgets("gives landscape value when device landscape",
           (WidgetTester tester) async {
-        const portraitText = "portrait";
+        const originalText = "portrait";
         const landscapeText = "landscape";
-        tester.setDeviceSize(Size(800, 320));
+        tester.setDeviceSize(testLandscapeSize);
         await tester.pumpWidget(Builder(builder: (context) {
-          return Text(portraitText.whenLandscape(context, landscapeText));
+          return Text(originalText.valueWhenOrientaitonLandscape(
+              context, landscapeText));
         }).wrapWithMaterialApp());
         final findLandscapeText = find.text(landscapeText);
         expect(findLandscapeText, findsOneWidget);
       });
+
+      testWidgets("gives original value when device portrait",
+          (WidgetTester tester) async {
+        const originalText = "portrait";
+        const landscapeText = "landscape";
+        tester.setDeviceSize(testPortraitSize);
+        await tester.pumpWidget(Builder(builder: (context) {
+          return Text(originalText.valueWhenOrientaitonLandscape(
+              context, landscapeText));
+        }).wrapWithMaterialApp());
+        final findLandscapeText = find.text(originalText);
+        expect(findLandscapeText, findsOneWidget);
+      });
     });
-    group("when orientation is portrait", () {
-      testWidgets("get portrait value when portrait",
+    group("valueWhenOrientationPortrait", () {
+      testWidgets("gives portrait value when orientation is portrait",
           (WidgetTester tester) async {
         const portraitText = "portrait";
-        const landscapeText = "landscape";
-        tester.setDeviceSize(Size(320, 800));
+        const originalText = "landscape";
+        tester.setDeviceSize(testPortraitSize);
         await tester.pumpWidget(Builder(builder: (context) {
-          return Text(landscapeText.whenPortrait(context, portraitText));
+          return Text(
+              originalText.valueWhenOrientationPortrait(context, portraitText));
         }).wrapWithMaterialApp());
         final findPortraitText = find.text(portraitText);
         expect(findPortraitText, findsOneWidget);
       });
+
+      testWidgets("gives original value when orientation is landscape",
+          (WidgetTester tester) async {
+        const portraitText = "portrait";
+        const originalText = "landscape";
+        tester.setDeviceSize(testLandscapeSize);
+        await tester.pumpWidget(Builder(builder: (context) {
+          return Text(
+              originalText.valueWhenOrientationPortrait(context, portraitText));
+        }).wrapWithMaterialApp());
+        final findPortraitText = find.text(originalText);
+        expect(findPortraitText, findsOneWidget);
+      });
     });
   });
+
+  group("ValueByOrientation", () {
+    final landScapeText = "landscape";
+    final portraitText = "portrait";
+
+    test("instantiates", () {
+      final valueByOrientation =
+          ValueByOrientation<int>(24, 45, DummyContext());
+      expect(valueByOrientation, isNotNull);
+    });
+
+    /// This is very difficult to mock mediaquery data with a Mock Buildcontext
+    /// Github Issue: [ct_test_helpers/1](https://github.com/cookytech/ct_test_helpers/issues/1)
+    /// we need a mock context for running unit tests. For now we can test these values
+    /// using a tester widget and running widget tests
+    testWidgets("returns landscape value in landscape mode",
+        (WidgetTester tester) async {
+      // pump a text widget with the required value and then test for the correct value
+      tester.setDeviceSize(testLandscapeSize);
+      await tester.pumpWidget(
+        ValueByOrientationTesterWidget(
+          landscapeText: landScapeText,
+          portraitText: portraitText,
+        ).wrapWithMaterialApp(),
+      );
+      final findLandscapeText = find.text(landScapeText);
+      expect(findLandscapeText, findsOneWidget);
+    });
+    testWidgets("returns portrait value in portrait mode",
+        (WidgetTester tester) async {
+      // pump a text widget with the required value and then test for the correct value
+      tester.setDeviceSize(testPortraitSize);
+      await tester.pumpWidget(
+        ValueByOrientationTesterWidget(
+          landscapeText: landScapeText,
+          portraitText: portraitText,
+        ).wrapWithMaterialApp(),
+      );
+      final findPortraitText = find.text(portraitText);
+      expect(findPortraitText, findsOneWidget);
+    });
+  });
+}
+
+class ValueByOrientationTesterWidget extends StatelessWidget {
+  final String landscapeText, portraitText;
+
+  const ValueByOrientationTesterWidget(
+      {Key key, this.landscapeText, this.portraitText})
+      : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      child:
+          Text(ValueByOrientation(landscapeText, portraitText, context).value),
+    );
+  }
 }