+++ CHANGELOG.md
@@ -1,4 +1,8 @@
-## [0.0.1] - 02.07.2020
+## [0.1.0+2] - 12.07.2020
+* ResponsiveByOrientation and extension functions
+* ValueByOrientation and extension functions
+
+## [0.0.1+1] - 02.07.2020
 
 * Initial release with readme description
 * Empty project, should not be used.